+++ src/Petabridge.Tracing.Zipkin.Tests/Serialization/JsonSerializationSpecs.cs
@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="JsonSerializationSpecs.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2018 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -115,7 +115,8 @@ namespace Petabridge.Tracing.Zipkin.Tests.Serialization
             var expectedBytes = Encoding.UTF8.GetBytes(json);
 
             var span = new Span(Tracer, "op1",
-                    new SpanContext(new TraceId(7776525154056436086, 6707114971141086261), -7118946577185884628, null,
+                    new SpanContext(new TraceId(7776525154056436086, 6707114971141086261),
+                        (-7118946577185884628).ToString("x16"), null,
                         true),
                     startTime, SpanKind.CLIENT)
                 .SetRemoteEndpoint(new Endpoint("actorsystem", "127.0.0.1", 8009)).SetTag("foo1", "bar")
@@ -129,8 +130,9 @@ namespace Petabridge.Tracing.Zipkin.Tests.Serialization
         [Fact(
             DisplayName =
                 "Should be able to serialize a span with a defined parent into a valid Zipkin-friendly JSON format.",
-            Skip = "Weird environmental stuff on build server.")]
-        public void ShouldMapSpanWithparentIntoValidJson()
+            Skip = "Weird environmental stuff on build server."
+        )]
+        public void ShouldMapSpanWithParentIntoValidJson()
         {
             var json = @"
                 [
@@ -176,7 +178,8 @@ namespace Petabridge.Tracing.Zipkin.Tests.Serialization
             var parentId = 11210001.ToString("x16");
 
             var span = new Span(Tracer, "op1",
-                    new SpanContext(new TraceId(7776525154056436086, 6707114971141086261), -7118946577185884628,
+                    new SpanContext(new TraceId(7776525154056436086, 6707114971141086261),
+                        (-7118946577185884628).ToString("x16"),
                         parentId,
                         true),
                     startTime, SpanKind.CLIENT)