+++ src/Petabridge.Tracing.Zipkin.Tests/SpanContextSpecs.cs
@@ -1,6 +1,6 @@
 ﻿// -----------------------------------------------------------------------
 // <copyright file="SpanContextSpecs.cs" company="Petabridge, LLC">
-//      Copyright (C) 2018 - 2018 Petabridge, LLC <https://petabridge.com>
+//      Copyright (C) 2015 - 2018 Petabridge, LLC <https://petabridge.com>
 // </copyright>
 // -----------------------------------------------------------------------
 
@@ -15,7 +15,7 @@ namespace Petabridge.Tracing.Zipkin.Tests
         public void SpanContextSampledShouldAlwaysBeFalseWhenDebugEnabled()
         {
             // explicitly set both DEBUG and SAMPLED to true
-            var spanContext = new SpanContext(new TraceId(90, 0), 1, null, true, true);
+            var spanContext = new SpanContext(new TraceId(90, 0), "1", null, true, true);
             spanContext.Debug.Should().BeTrue();
             spanContext.Sampled.Should().BeFalse();
         }